:recycle: logger function out to protocol
